more linemod clean-up: fixed some compile errors and warnings, did a bit more tidying
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@4844 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ recognition/include/pcl/recognition/sparse_quantized_multi_mod_template.h
@@ -49,7 +49,7 @@ namespace pcl
   {
     int x;
     int y;
-    int modality_index;
+    size_t modality_index;
     unsigned char quantized_value;
 
     void
